Let events bubble up, don't cancel in PlayerInteractEvent
There is an issue where PlayerInteractEvent is called before BlockBreakEvent so breaking a chest would cause odd behavior like allowing the shop creation dialog if there is an existing shop. Additionally, since it cancels the event while its a PlayerInteractionEvent it never becomes a BlockBreakEvent and so you are unable to remove shops in some cases.