Safe file updates using Atomic writes and backups!
Previously we deleted the old player save file before creating the yaml, so if there was an error thrown generating the yaml it would cause the file to be deleted and the new file to never be written. This was leading to the reports from users that all of their shop files would be deleted on startup, since if there was invalid NBT data for example, the file would be deleted but would not be able to save the new file.

This new method adds atomic file moves/writes to allow it to be very safe, and if ATOMIC_MOVE is not supported or there is some error, it will attempt to save it by moving the current player file to a backup, swapping in the new file (written to a tmp location), then deleting the backup if that was successful. It will restore the backup upon error, and also notify the user of any issues that occur.

Might be related to:
snowgears/shopbugs#458